# Changed: AccumuloGraph to use MultiTableBatchWriter